seperate branch from company and variant from item backend & frontend

+++ Modules/Branch/app/Http/Controllers/BranchDestroy.php
@@ -12,6 +12,7 @@ class BranchDestroy extends Controller
     public function __invoke(Branch $branch)
     {
         if (!auth()->user()->is_owner)  abort_if(Gate::denies('delete-branch'), Response::HTTP_FORBIDDEN, __('permission::messages.gate_denies'));
+        if($branch->is_main) return $this->error(__('status.delete_main_branch_error'), Response::HTTP_INTERNAL_SERVER_ERROR);
         try {
             $branch->delete();
             return $this->success(__('status.deleted', ['name' => $branch->name, 'module' => __('modules.branch')]));
